Refactor user registration and authentication routes. Moved registration route to app directory, added user routes, and updated user creation logic to accept Clerk user object for enhanced security and clarity.
